+++ src/Core/GUI/SwingEX/XLabel.java
@@ -1,4 +1,6 @@
-package Core.GUI;
+package Core.GUI.SwingEX;
+
+import Core.GUI.gfxRepository;
 
 import javax.swing.*;
 import java.awt.*;
@@ -10,16 +12,16 @@ import java.awt.image.BufferedImage;
  * Extends JLabel with extra functionality and constructors.
  * Essentially simplifies Swing object declaration so I don't have to use as many lines.
  */
-class XLabel extends JLabel {
+public class XLabel extends JLabel {
 
     //default constructor
-    XLabel() {
+    public XLabel() {
         this.setFocusable(false);
 
     }
 
     //animation
-    XLabel(Icon animation) {
+    public XLabel(Icon animation) {
         this.setIcon(animation);
         this.setOpaque(false);
         this.setFocusable(false);
@@ -28,14 +30,14 @@ class XLabel extends JLabel {
     }
 
     //text
-    XLabel(String text, Font font, Color colour) {
+    public XLabel(String text, Font font, Color colour) {
         this.setText(text, font, colour);
         this.setOpaque(false);
 
     }
 
     //text without text
-    XLabel(Font font, Color textColour) {
+    public XLabel(Font font, Color textColour) {
         this.setFont(font);
         this.setForeground(textColour);
         this.setOpaque(false);
@@ -43,7 +45,7 @@ class XLabel extends JLabel {
     }
 
     //images
-    XLabel(BufferedImage image) {
+    public XLabel(BufferedImage image) {
         this.setIcon(new ImageIcon(image));
         this.setBackground(null); //transparent background
         this.setBorder(null);
@@ -54,7 +56,7 @@ class XLabel extends JLabel {
     }
 
     //images with background
-    XLabel(BufferedImage image, Color colour) {
+    public XLabel(BufferedImage image, Color colour) {
         this.setIcon(new ImageIcon(image));
         this.setBackground(colour); //if the image doesn't work, just render a black background
         this.setBorder(null);
@@ -65,7 +67,7 @@ class XLabel extends JLabel {
 
     //in the event that the default constructor is used, these methods will set all of the data
 
-    void setText(String text, Font font, Color colour) {
+    public void setText(String text, Font font, Color colour) {
         this.setText(text);
         this.setFont(font);
         this.setForeground(colour);
@@ -73,7 +75,7 @@ class XLabel extends JLabel {
 
     }
 
-    void setImage(BufferedImage image) {
+    public void setImage(BufferedImage image) {
         this.setIcon(new ImageIcon(image));
         this.setBackground(gfxRepository.clrTrueBlack);
         this.setBorder(null);
@@ -82,14 +84,14 @@ class XLabel extends JLabel {
     }
 
     //sets the horizontal and vertical alignment separately
-    void setAlignments(int alignmentH, int alignmentV) {
+    public void setAlignments(int alignmentH, int alignmentV) {
         this.setHorizontalAlignment(alignmentH);
         this.setVerticalAlignment(alignmentV);
 
     }
 
     //sets both the h and v alignment to the same value
-    void setAlignments(int alignment) {
+    public void setAlignments(int alignment) {
         this.setHorizontalAlignment(alignment);
         this.setVerticalAlignment(alignment);