+++ src/Core/GUI/SwingEX/XToolTipUI.java
@@ -1,0 +1,17 @@
+package Core.GUI.SwingEX;
+
+import javax.swing.plaf.basic.BasicToolTipUI;
+
+/**
+ * KM
+ * May 26 2017
+ * Rewrites the ToolTipUI's design.
+ */
+
+public class XToolTipUI extends BasicToolTipUI {
+
+        public XToolTipUI() {
+
+        }
+
+}