+++ src/Core/Player/DataConstants.java
@@ -1,0 +1,15 @@
+package Core.Player;
+
+/**
+ * KM
+ * May 26 2017
+ * Constants for data initialization, used by the saveData class.
+ */
+
+public interface DataConstants {
+    int FOLDER = 1;
+    int SERIAL = 2;
+    int TEXT = 3;
+    int IMAGE = 4;
+    int XML = 5;
+}